Stitch animation frames together (#5)
Compiling art is cool and all, but it'd be even cooler if we could share the video.  As a step toward #3, this adds a way to run the animation for a fixed number of frames and output those frames to PNGs.  There's also an `ffmpeg` recipe for stitching these together into a video.

Still TODO:
* Allow CLI args to set the animation time
* Stop relying on `ffmpeg` to pack the frames together.