Use CLI args to create model (#6)
Continuing from #5 and on our way to #3, this threads the CLI args through to the model.  (What I was missing in that late-night PR was that we could parse the options _in_ the model function.)

This makes frame output a "mode": it only writes the frame PNGs if `--output-dir` is set.